This commit updates the parmap calls to use the new parmap arguments.
The last parmap version allows for keyword arguments to be passed to
parallelized functions like this:

    parmap.map(yourfunc, your_list, your_arg=your_val)

In order to prevent confusion between parmap arguments and keyword
arguments passed to yourfunc, it was necessary to use a common prefix
for all parmap parameters.

    parmap.map(yourfunc, your_list, your_arg=your_val, pm_parallel=True)

With this new syntax, it is easy for anyone to see that keyword arguments
starting with `pm_` are for parmap, while other arguments (here your_arg)
will be passed to the parallelized function.

That's why I had to change `parallel` to `pm_parallel`.

On the other hand, the best way to use parmap with a specific number
of CPUs is through the `pm_processes` argument, instead of creating
a multiprocessing pool manually. This has been changed in this commit
as well.

Currently the old parmap arguments are still valid, and they give
deprecation warnings so users can update the syntax. However it is best
to transition to the new arguments, as a future parmap 2.0 version
might break this backwards compatibility.

The parmap dependency in setup.py has been updated accordingly as well.